pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "litecharts"
version = "0.1.0"
description = "Python wrapper for TradingView Lightweight Charts"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Chad Thackray" }
]
keywords = ["trading", "charts", "visualization", "finance", "candlestick"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Visualization",
    "Typing :: Typed",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "mypy>=1.0",
    "ruff>=0.8",
    "pandas-stubs>=2.0",
]

[project.urls]
Homepage = "https://github.com/chad/litecharts"
Repository = "https://github.com/chad/litecharts"

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/build_hooks.py",
]

[tool.hatch.build.targets.wheel]
packages = ["src/litecharts"]

[tool.hatch.build.hooks.custom]
path = "build_hooks.py"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.ruff]
target-version = "py310"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF", "N"]
ignore = [
    "N802",  # Function name should be lowercase (we use camelCase to match LWC)
    "N803",  # Argument name should be lowercase (we use camelCase to match LWC)
    "N806",  # Variable in function should be lowercase (we use camelCase to match LWC)
    "N815",  # Variable in class scope should not be mixedCase (we use camelCase)
    "N816",  # Variable in global scope should not be mixedCase (we use camelCase)
]
